renamed and moved a couple files to better reflect what I want
